Increased heartbeat timeout during automated testing.

+++ Providers/Common/ProviderUtil/FIX/FIXProviderSupport.cs
@@ -333,8 +333,8 @@ namespace TickZoom.FIX
 		}
 
 		protected void IncreaseRetryTimeout() {
-			retryTimeout = Factory.Parallel.TickCount + retryDelay * 1000;
-			heartbeatTimeout = Factory.Parallel.TickCount + heartbeatDelay * 1000;
+			retryTimeout = Factory.Parallel.TickCount + retryDelay * 1000L;
+			heartbeatTimeout = Factory.Parallel.TickCount + (long)heartbeatDelay * 1000L;
 		}
 		
 		protected abstract void OnStartRecovery();
